requirements.txt
fastapi
uvicorn
python-dotenv
google-cloud-vision
requests
pydantic
python-multipart